# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for scheduling and managing messages",
        "title": "Go Message Scheduler API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "muratgun545@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/sent-messages": {
            "get": {
                "description": "Get all successfully sent messages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Retrieve all sent messages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Message"
                            }
                        }
                    },
                    "404": {
                        "description": "No sent messages found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/worker-pool/state": {
            "put": {
                "description": "Start or pause the worker pool",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "worker-pool"
                ],
                "summary": "Updates the worker pool state",
                "parameters": [
                    {
                        "description": "Action to perform `start` or `pause`",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.WorkerPoolActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.WorkerPoolStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid action",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recipient_phone_number": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "webhook_response_message_id": {
                    "type": "string"
                }
            }
        },
        "main.WorkerPoolActionRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "\"start\" or \"pause\"",
                    "type": "string"
                }
            }
        },
        "main.WorkerPoolStatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}